+++ src/Ingredients/RelationIngredient.php
@@ -2,12 +2,13 @@
 
 namespace Serri\Alchemist\Ingredients;
 
+use Serri\Alchemist\Contracts\AcceptsNestedFormula;
 use Serri\Alchemist\Contracts\IngredientContract;
 use Serri\Alchemist\Decorators\Relation;
 use Serri\Alchemist\Helpers\DecoratorHelper;
 use Serri\Alchemist\Services\Alchemist;
 
-final class RelationIngredient implements IngredientContract
+final class RelationIngredient implements AcceptsNestedFormula, IngredientContract
 {
     public static function usesDecorator(): bool
     {
@@ -25,6 +26,15 @@ final class RelationIngredient implements IngredientContract
 
         return [
             $ingredient => app(Alchemist::class)->brew($brewing->$relationName),
+        ];
+    }
+
+    public static function infuseWithFormula(string $ingredient, mixed $brewing, array $formula): array
+    {
+        $relationName = self::getRelationName($brewing, $ingredient);
+
+        return [
+            $ingredient => app(Alchemist::class)->brew($brewing->$relationName, $formula),
         ];
     }